refactor/EV-804/componetização finalizada
