+++ lib/main.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 
-import 'coins/coins_pages.dart';
+import 'presenter/coins/coins_pages.dart';
 
 void main() {
   runApp(ProviderScope(child: const MyApp()));
