+++ lib/presenter/coins/widgets/body_coins.dart
@@ -2,8 +2,8 @@ import 'package:flutter/material.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
 import 'package:intl/intl.dart';
 
-import '../provider/provider.dart';
-import '../repositories/coins_repositry.dart';
+import '../../../shared/provides/provider.dart';
+import '../../../repositories/coins_repositry.dart';
 import 'list_coins.dart';
 
 class BodyCoins extends ConsumerStatefulWidget {
